+++ src/components/datagrid/datagrid.formatters.js
@@ -689,7 +689,6 @@ const formatters = {
   // Process Indicator
   // File Upload (Simple)
   // Menu Button
-  // Color Picker (Low)
   // Radio
 };