+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### v4.21.0 Features
 
+- `[Datagrid]` Added the ability to use frozen columns with tree grid. ([#'2102](https://github.com/infor-design/enterprise/issues/'2102))
 - `[Homepage]` Added a parameter to the `resize` event that provides metadata about the Homepage's state, including a calculated container height. ([#2446](https://github.com/infor-design/enterprise/issues/2446))
 
 ### v4.21.0 Fixes